fix: minor review comments
