+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.constants.js
@@ -12,3 +12,10 @@ export const enrollmentAccessLevels = {
     NO_ACCESS: 'NO_ACCESS',
     UNKNOWN_ACCESS: 'UNKNOWN_ACCESS',
 };
+
+export const serverErrorMessages = {
+    OWNERSHIP_ACCESS_PARTIALLY_DENIED: 'OWNERSHIP_ACCESS_PARTIALLY_DENIED',
+    OWNERSHIP_ACCESS_DENIED: 'OWNERSHIP_ACCESS_DENIED',
+    PROGRAM_ACCESS_CLOSED: 'PROGRAM_ACCESS_CLOSED',
+    ORGUNIT_OUT_OF_SCOPE: '[User has no read access to organisation unit:',
+};
