+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.epics.js
@@ -15,10 +15,9 @@ import {
     startFetchingTeiFromEnrollmentId,
     startFetchingTeiFromTeiId,
 } from './EnrollmentPage.actions';
-import { enrollmentAccessLevels } from './EnrollmentPage.constants';
+import { enrollmentAccessLevels, serverErrorMessages } from './EnrollmentPage.constants';
 import { buildUrlQueryString, getLocationQuery } from '../../../utils/routing';
 import { deriveTeiName } from '../common/EnrollmentOverviewDomain/useTeiDisplayName';
-import { serverErrorMessages } from '../../../constants';
 
 const sortByDate = (enrollments = []) => enrollments.sort((a, b) =>
     moment.utc(b.enrolledAt).diff(moment.utc(a.enrolledAt)));
