+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.types.js
@@ -12,7 +12,7 @@ export type ContainerProps = $ReadOnly<{|
   trackedEntityName: string,
   teiDisplayName: string,
   enrollmentsAsOptions: Array<Object>,
-  selectProgramHandler: () => void,
+  selectProgramHandler: (id: string) => void,
   deselectProgramHandler: () => void,
 |}>
