+++ src/core_modules/capture-core/components/WidgetBreakingTheGlass/WidgetBreakingTheGlass.types.js
@@ -3,12 +3,12 @@
 export type Props = {|
     programId: string,
     teiId: string,
-    onBreakingTheGlass: Function,
-    onCancel: Function,
+    onBreakingTheGlass: (?string) => void,
+    onCancel: () => void,
 |};
 
 export type PlainProps = {|
-    onBreakingTheGlass: Function,
-    onCancel: Function,
+    onBreakingTheGlass: (?string) => void,
+    onCancel: () => void,
     ...CssClasses,
 |};
