+++ src/core_modules/capture-core/constants/index.js
@@ -3,4 +3,3 @@ export { LOADING_INDICATOR_TIMEOUT } from './loadingIndicatorTimeout';
 export { environments } from './environments';
 export { FEATURETYPE } from './geometry';
 export { dataEntryIds, dataEntryKeys } from './dataEntries';
-export { serverErrorMessages } from './serverErrorMessages';
