--- src/core_modules/capture-core/constants/serverErrorMessages.js
@@ -1,8 +1,0 @@
-// @flow
-
-export const serverErrorMessages = {
-    OWNERSHIP_ACCESS_PARTIALLY_DENIED: 'OWNERSHIP_ACCESS_PARTIALLY_DENIED',
-    OWNERSHIP_ACCESS_DENIED: 'OWNERSHIP_ACCESS_DENIED',
-    PROGRAM_ACCESS_CLOSED: 'PROGRAM_ACCESS_CLOSED',
-    ORGUNIT_OUT_OF_SCOPE: '[User has no read access to organisation unit:',
-};
